refactor(android): move `findOutputFile` fn
